Modify response types

+++ app/src/main/java/com/globallogic/thespaceapp/data/remote/response/rockets/Diameter.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Diameter(
     @SerializedName("feet")
-    val feet: Int,
+    val feet: Double,
     @SerializedName("meters")
-    val meters: Int
+    val meters: Double
 )
