+++ app/src/main/java/com/globallogic/thespaceapp/data/remote/response/rockets/PayloadWeight.kt
@@ -7,9 +7,9 @@ data class PayloadWeight(
     @SerializedName("id")
     val id: String,
     @SerializedName("kg")
-    val kg: Int,
+    val kg: Double,
     @SerializedName("lb")
-    val lb: Int,
+    val lb: Double,
     @SerializedName("name")
     val name: String
 )
