+++ app/src/main/java/com/globallogic/thespaceapp/data/remote/response/rockets/ThrustSeaLevel.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ThrustSeaLevel(
     @SerializedName("kN")
-    val kN: Int,
+    val kN: Double,
     @SerializedName("lbf")
-    val lbf: Int
+    val lbf: Double
 )
